Added stan code which had been missing before. Added ofc_learning_choosing_path.m, which we'll ask users to edit to give the relative path and use to create absolute paths. Modified code so that driver_1.m is now able to fit models.

+++ Library/behavior_analysis/ratdata2standata.m
@@ -4,11 +4,12 @@ function standata = ratdata2standata(ratdata,inc)
 
     standata.pCong = 0.6*round(ratdata.p_congruent) + 0.2; % Round to exactly 0.2 or 0.8
     standata.nTrials = ratdata.nTrials;
-    standata.choices = (ratdata.sides1=='l')+1;
-    standata.outcomes = (ratdata.sides2=='l')+1;
-    standata.commons = ratdata.trans_common;
-
-    standata.rewards = ratdata.rewards;
+    % Convert to int8 since Stan requires a numerical data type, rather
+    % than a logical
+    standata.choices = int8((ratdata.sides1=='l')+1);
+    standata.outcomes = int8((ratdata.sides2=='l')+1);
+    standata.commons = int8(ratdata.trans_common);
+    standata.rewards = int8(ratdata.rewards);
     
     if exist('inc','var')
